refactor(agent-loop): split quality continuations
